[Refactor] Refactor InfoStringFactory
